+++ libs/Python/pybind/visualization/gui/gui.h
@@ -21,11 +21,14 @@ class CloudViewerScene;
 
 namespace gui {
 
-void InitializeForPython(std::string resource_path = "");
+void InitializeForPython(std::string resource_path = "", bool headless = false);
 std::shared_ptr<geometry::Image> RenderToImageWithoutWindow(
         rendering::CloudViewerScene *scene, int width, int height);
 std::shared_ptr<geometry::Image> RenderToDepthImageWithoutWindow(
-        rendering::CloudViewerScene *scene, int width, int height);
+        rendering::CloudViewerScene *scene,
+        int width,
+        int height,
+        bool z_in_view_space = false);
 
 void pybind_gui(py::module &m);