+++ libs/cloudViewer/geometry/VoxelGrid.h
@@ -4,3 +4,5 @@
 // Maps cloudViewer/geometry/VoxelGrid.h -> repo's VoxelGrid.h (eCV_db)
 
 #include <VoxelGrid.h>
+
+#include "cloudViewer/geometry/Geometry.h"