+++ libs/cloudViewer/visualization/visualizer/GuiSettingsModel.cpp
@@ -157,15 +157,8 @@ GuiSettingsModel::GetDefaultLightingProfile() {
 
 const GuiSettingsModel::LightingProfile&
 GuiSettingsModel::GetDefaultPointCloudLightingProfile() {
-    for (auto& lp : GuiSettingsModel::lighting_profiles_) {
-        if (lp.name == POINT_CLOUD_PROFILE_NAME) {
-            return lp;
-        }
-    }
-    utility::LogWarning(
-            "Internal Error: could not find default point cloud lighting "
-            "profile");
-    return GuiSettingsModel::GetDefaultLightingProfile();
+    // Lighting profile 0 will always be default even for Point Clouds
+    return GuiSettingsModel::lighting_profiles_[0];
 }
 
 const GuiSettingsModel::LitMaterial& GuiSettingsModel::GetDefaultLitMaterial() {
@@ -205,10 +198,23 @@ void GuiSettingsModel::SetShowGround(bool show) {
     NotifyChanged();
 }
 
-
 bool GuiSettingsModel::GetSunFollowsCamera() const { return sun_follows_cam_; }
 void GuiSettingsModel::SetSunFollowsCamera(bool follow) {
     sun_follows_cam_ = follow;
+    NotifyChanged();
+}
+
+bool GuiSettingsModel::GetBasicMode() const { return basic_mode_enabled_; }
+void GuiSettingsModel::SetBasicMode(bool enable) {
+    basic_mode_enabled_ = enable;
+    NotifyChanged(true);
+}
+
+bool GuiSettingsModel::GetWireframeMode() const {
+    return wireframe_mode_enabled_;
+}
+void GuiSettingsModel::SetWireframeMode(bool enable) {
+    wireframe_mode_enabled_ = enable;
     NotifyChanged();
 }
 
@@ -223,6 +229,7 @@ void GuiSettingsModel::SetBackgroundColor(const Eigen::Vector3f& color) {
 const GuiSettingsModel::LightingProfile& GuiSettingsModel::GetLighting() const {
     return lighting_;
 }
+
 void GuiSettingsModel::SetLightingProfile(const LightingProfile& profile) {
     lighting_ = profile;
     user_has_changed_lighting_profile_ = true;
@@ -329,7 +336,7 @@ void GuiSettingsModel::ResetColors() {
     NotifyChanged(true);
 }
 
-void GuiSettingsModel::SetCustomDefaultColor(const Eigen::Vector3f& color) {
+void GuiSettingsModel::SetCustomDefaultColor(const Eigen::Vector3f color) {
     custom_default_color = color;
 }
 
@@ -344,6 +351,20 @@ int GuiSettingsModel::GetPointSize() const {
 void GuiSettingsModel::SetPointSize(int size) {
     current_materials_.point_size = float(size);
     NotifyChanged(true);
+}
+
+bool GuiSettingsModel::GetUserWantsEstimateNormals() {
+    if (user_wants_estimate_normals_) {
+        user_wants_estimate_normals_ = false;
+        return true;
+    } else {
+        return false;
+    }
+}
+
+void GuiSettingsModel::EstimateNormalsClicked() {
+    user_wants_estimate_normals_ = true;
+    NotifyChanged();
 }
 
 bool GuiSettingsModel::GetDisplayingPointClouds() const {