+++ libs/cloudViewer/visualization/visualizer/GuiSettingsView.h
@@ -37,6 +37,7 @@ public:
                     std::function<void(const char*)> on_load_ibl);
 
     void ShowFileMaterialEntry(bool show);
+    void EnableEstimateNormals(bool enable);
     void Update();
 
 private:
@@ -64,6 +65,12 @@ private:
     std::shared_ptr<gui::ColorEdit> material_color_;
     std::shared_ptr<gui::Button> reset_material_color_;
     std::shared_ptr<gui::Slider> point_size_;
+    std::shared_ptr<gui::Button> generate_normals_;
+    std::shared_ptr<gui::Checkbox> basic_mode_;
+    std::shared_ptr<gui::Checkbox> wireframe_mode_;
+
+    bool sun_follows_cam_was_on_ = true;
+    void UpdateUIForBasicMode(bool enable);
 };
 
 }  // namespace visualization