+++ libs/cloudViewer/visualization/visualizer/O3DVisualizer.h
@@ -60,7 +60,7 @@ public:
                 gui::SceneWidget::Controls::ROTATE_CAMERA;
         Shader scene_shader = Shader::STANDARD;
         bool show_settings = false;
-        bool show_skybox = false;
+        bool show_skybox = true;
         bool show_axes = false;
         bool show_ground = false;
         rendering::Scene::GroundPlane ground_plane =
@@ -74,6 +74,7 @@ public:
 
         bool use_ibl = false;
         bool use_sun = true;
+        bool sun_follows_camera = true;
         std::string ibl_path = "";  // "" is default path
         int ibl_intensity = 0;
         int sun_intensity = 100000;
@@ -117,11 +118,25 @@ public:
                      double time = 0.0,
                      bool is_visible = true);
 
+    /// Removes the named geometry from the Visualizer
     void RemoveGeometry(const std::string& name);
 
+    /// Updates `update_flags` attributes of named geometry with the matching
+    /// attributes from `tgeom`
+    void UpdateGeometry(const std::string& name,
+                        std::shared_ptr<t::geometry::Geometry> tgeom,
+                        uint32_t update_flags);
+
+    /// Show/hide the named geometry
     void ShowGeometry(const std::string& name, bool show);
 
+    /// Returns Visualizer's internal DrawObject for the named geometry
     DrawObject GetGeometry(const std::string& name) const;
+    rendering::MaterialRecord GetGeometryMaterial(
+            const std::string& name) const;
+
+    void ModifyGeometryMaterial(const std::string& name,
+                                const rendering::MaterialRecord* material);
 
     void Add3DLabel(const Eigen::Vector3f& pos, const char* text);
     void Clear3DLabels();
@@ -141,13 +156,19 @@ public:
 
     void ShowSettings(bool show);
     void ShowSkybox(bool show);
+    void SetIBL(const std::string& path);
+    void SetIBLIntensity(float intensity);
     void ShowAxes(bool show);
     void ShowGround(bool show);
     void SetGroundPlane(rendering::Scene::GroundPlane plane);
+    void EnableSunFollowsCamera(bool enable);
+    void EnableBasicMode(bool enable);
+    void EnableWireframeMode(bool enable);
     void SetPointSize(int point_size);
     void SetLineWidth(int line_width);
     void EnableGroup(const std::string& group, bool enable);
     void SetMouseMode(gui::SceneWidget::Controls mode);
+    void SetPanelOpen(const std::string& name, bool open);
 
     std::vector<O3DVisualizerSelections::SelectionSet> GetSelectionSets() const;
 
@@ -177,7 +198,7 @@ public:
     UIState GetUIState() const;
     rendering::CloudViewerScene* GetScene() const;
 
-    /// Starts the RPC interface. See io/rpc/ReceiverBase for the parameters.
+    /// Starts the RPC interface. See io/rpc/ZMQReceiver for the parameters.
     void StartRPCInterface(const std::string& address, int timeout);
 
     void StopRPCInterface();